// frontend/src/app/Models/person.model.ts
export class Person {
    
    id:             number;
    user:           string;
    pass:           string;
    name:           string;
    lastName:       string;
    documentType:   string;
    documentNumber: string;
    cuit:           string;
    cuil:           string;
    socialReason:   string; 
    city:           string;
    street:         string;
    phoneNumber_1:  string;
    phoneNumber_2:  string;
    celPhone_1:     string;
    celPhone_2:     string;
    email_1:        string;
    email_2:        string;
    highDate:       string;
    status:         boolean;
    userType:       number
}